+++ CHANGELOG.md
@@ -23,6 +23,7 @@
 - `proxy` mapping, for enabling or disabling the use of proxies when using guest tokens
 - Caching IDs of posts published and skip mirroring tweet if associated published post is found to avoid duplicates. You can change this with the `avoid_duplicates` mapping.
 - `application_name` mapping, for specifying the Fediverse application name to use as a filter when getting the date of the latest published post by the bot. Thanks @reorx!
+- Configuration wizard, which generates a minimal config if none is found
 
 ## [1.1.0] 06-06-2022
 ## Fixed